hms_backend/requirements.txt
asgiref==3.11.0
brotli==1.2.0
cffi==2.0.0
cryptography==46.0.3
cssselect2==0.8.0
dj-database-url==3.0.1
Django==6.0
django-appconf==1.2.0
django-cryptography==1.1
django-extensions==4.1
django-filter==25.2
django-simple-history==3.11.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
fonttools==4.61.1
gunicorn==23.0.0
packaging==25.0
pillow==12.1.0
psycopg2-binary==2.9.11
pycparser==2.23
pydyf==0.12.1
PyJWT==2.10.1
pyphen==0.17.2
python-decouple==3.8
sqlparse==0.5.5
tinycss2==1.5.1
tinyhtml5==2.0.0
weasyprint==68.1
webencodings==0.5.1
whitenoise==6.12.0
zopfli==0.4.0
